+++ weblate/addons/models.py
@@ -190,6 +190,7 @@ class AddonsConf(AppConf):
         "weblate.addons.generate.GenerateFileAddon",
         "weblate.addons.generate.PseudolocaleAddon",
         "weblate.addons.generate.PrefillAddon",
+        "weblate.addons.generate.FillReadOnlyAddon",
         "weblate.addons.json.JSONCustomizeAddon",
         "weblate.addons.xml.XMLCustomizeAddon",
         "weblate.addons.properties.PropertiesSortAddon",